Move exceptions to own package
