+++ caching-service/src/main/java/org/zowe/apiml/caching/service/redis/RedisOperator.java
@@ -22,6 +22,9 @@ import lombok.NoArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import org.zowe.apiml.caching.model.KeyValue;
+import org.zowe.apiml.caching.service.redis.exceptions.RedisEntryException;
+import org.zowe.apiml.caching.service.redis.exceptions.RedisOutOfMemoryException;
+import org.zowe.apiml.caching.service.redis.exceptions.RetryableRedisException;
 import org.zowe.apiml.message.log.ApimlLogger;
 
 import javax.annotation.PreDestroy;
