+++ caching-service/src/main/java/org/zowe/apiml/caching/service/redis/exceptions/RedisOutOfMemoryException.java
@@ -7,7 +7,7 @@
  *
  * Copyright Contributors to the Zowe Project.
  */
-package org.zowe.apiml.caching.service.redis;
+package org.zowe.apiml.caching.service.redis.exceptions;
 
 public class RedisOutOfMemoryException extends Exception {
     public RedisOutOfMemoryException(Throwable cause) {
