+++ caching-service/src/main/java/org/zowe/apiml/caching/service/redis/exceptions/RetryableRedisException.java
@@ -7,14 +7,10 @@
  *
  * Copyright Contributors to the Zowe Project.
  */
-package org.zowe.apiml.caching.service.redis;
+package org.zowe.apiml.caching.service.redis.exceptions;
 
-public class RedisEntryException extends Exception {
-    public RedisEntryException(String message) {
-        super(message);
-    }
-
-    public RedisEntryException(String message, Throwable cause) {
-        super(message, cause);
+public class RetryableRedisException extends RuntimeException {
+    public RetryableRedisException(Throwable e) {
+        super(e);
     }
 }
